Fixes to refactor
